+++ app/src/main/java/joshuatee/wx/ui/UtilityUI.kt
@@ -24,7 +24,6 @@ package joshuatee.wx.ui
 import android.app.Activity
 import android.content.Context
 import android.content.res.Configuration
-import android.graphics.Color
 import android.util.TypedValue
 import android.view.View
 import android.widget.LinearLayout
@@ -32,7 +31,6 @@ import android.widget.Toast
 import android.widget.TextView
 import kotlin.math.sqrt
 import kotlin.math.pow
-import com.google.android.material.snackbar.Snackbar
 import joshuatee.wx.R
 import joshuatee.wx.MyApplication
 import joshuatee.wx.UIPreferences
@@ -49,20 +47,6 @@ object UtilityUI {
         toast.duration = Toast.LENGTH_LONG
         toast.view = linearLayout
         toast.show()
-    }
-
-    fun makeSnackBar(view: View, message: String) {
-        val snack = Snackbar.make(view, message, Snackbar.LENGTH_INDEFINITE)
-        snack.setActionTextColor(Color.YELLOW)
-        snack.setAction("DISMISS") { snack.dismiss() }
-        val viewSnack = snack.view
-        val textView = viewSnack.findViewById(com.google.android.material.R.id.snackbar_text) as TextView
-        val fgColor = Color.WHITE
-        val bgColor = Color.BLACK
-        textView.setTextColor(fgColor)
-        textView.setTextSize(TypedValue.COMPLEX_UNIT_PX, MyApplication.textSizeNormal)
-        viewSnack.setBackgroundColor(bgColor)
-        snack.show()
     }
 
     fun immersiveMode(activity: Activity) {